+++ src/lexer.l.cpp
@@ -160,7 +160,7 @@ include          BEGIN(INCLUSION);
 
 {identifier}                           { yylval.identifier = strdup(yytext);    return IDENT; }
 
--?[0-9]+                               { yylval.integer = atoi(yytext);         return INTEGER; }
+-?[0-9]+                               { yylval.integer = atol(yytext);         return INTEGER; }
 -?([0-9]+|([0-9]*\.[0-9]+){exponent}?) { yylval.real    = atof(yytext);         return REAL; }
 "'"                                    { yylval.byte    = h_handle_char('\'');  return CHAR; }
 "\""                                   { yylval.string  = h_handle_string('"'); return STRING; }
